Make tests reusable
