+++ tests/DataObjectGetStaticReturnTypeExtensionTest.php
@@ -23,7 +23,7 @@ use Injector;
 use MySQLDatabase;
 use MySQLPDODatabase;
 
-final class DataObjectGetStaticReturnTypeExtensionTest extends \PHPStan\Testing\TestCase
+final class DataObjectGetStaticReturnTypeExtensionTest extends ResolverTest
 {
     /** @var \PHPStan\Broker\Broker */
     private $broker;
@@ -77,10 +77,10 @@ final class DataObjectGetStaticReturnTypeExtensionTest extends \PHPStan\Testing\
                 sprintf('%s::inst()->get(%s::class)', Injector::class, MySQLPDODatabase::class),
             ],
             // 
-            [
+            /*[
                 sprintf('%s', Object::class),
                 sprintf('class ExtensionTest extends \DataExtension { public function Test() { return $this->getOwner(); } }'),
-            ]
+            ]*/
         ];
     }
 
@@ -150,140 +150,6 @@ final class DataObjectGetStaticReturnTypeExtensionTest extends \PHPStan\Testing\
             [],
             [],
             $dynamicFunctionReturnTypeExtensions
-        );
-    }
-
-    private function assertTypes(
-        string $file,
-        string $description,
-        string $expression,
-        array $dynamicMethodReturnTypeExtensions = [],
-        array $dynamicStaticMethodReturnTypeExtensions = [],
-        array $dynamicFunctionReturnTypeExtensions = [],
-        string $evaluatedPointExpression = 'die;'
-    )
-    {
-
-        // NOTE(Jake): 2018-04-21
-        //
-        // If I don't do this, I get a class not found error.
-        //
-        // This is due to PHPStan using a classmap for loading files in its
-        // 'data' folder.
-        //
-        include_once($file);
-
-        // NOTE(Jake): 2018-04-21
-        //
-        // Taken from:
-        // - phpstan\tests\PHPStan\Analyser\NodeScopeResolverTest.php
-        //
-        $this->processFile($file, function (\PhpParser\Node $node, Scope $scope) use ($description, $expression, $evaluatedPointExpression) {
-            $printer = new \PhpParser\PrettyPrinter\Standard();
-            $printedNode = $printer->prettyPrint([$node]);
-            if ($printedNode === $evaluatedPointExpression) {
-                /** @var \PhpParser\Node\Expr $expressionNode */
-                $expressionNode = $this->getParser()->parseString(sprintf('<?php %s;', $expression))[0];
-                $type = $scope->getType($expressionNode);
-                $this->assertTypeDescribe(
-                    $description,
-                    $type->describe(),
-                    sprintf('%s at %s', $expression, $evaluatedPointExpression)
-                );
-            }
-        }, $dynamicMethodReturnTypeExtensions, $dynamicStaticMethodReturnTypeExtensions, $dynamicFunctionReturnTypeExtensions);
-    }
-
-    private function processFile(
-        string $file, 
-        \Closure $callback, 
-        array $dynamicMethodReturnTypeExtensions = [],
-        array $dynamicStaticMethodReturnTypeExtensions = [],
-        array $dynamicFunctionReturnTypeExtensions = []
-    )
-    {
-        // NOTE(Jake): 2018-04-21
-        //
-        // Taken from:
-        // - phpstan\tests\PHPStan\Analyser\NodeScopeResolverTest.php
-        //
-        $phpDocStringResolver = $this->getContainer()->getByType(PhpDocStringResolver::class);
-
-        $printer = new \PhpParser\PrettyPrinter\Standard();
-        $broker = $this->createBroker();
-
-        // NOTE(Jake): 2018-04-22
-        //
-        // Hack in DynamicFunctionReturnType support
-        //
-        if ($dynamicFunctionReturnTypeExtensions) {
-            $hack = $broker->getDynamicFunctionReturnTypeExtensions();
-            $hack = array_merge($hack, $dynamicFunctionReturnTypeExtensions);
-            foreach ($dynamicFunctionReturnTypeExtensions as $extension) {
-                if ($extension instanceof BrokerAwareExtension) {
-                    $extension->setBroker($broker);
-                }
-            }
-            $refProperty = new \ReflectionProperty($broker, 'dynamicFunctionReturnTypeExtensions');
-            $refProperty->setAccessible(true);
-            $refProperty->setValue($broker, $hack);
-        }
-
-        $resolver = new NodeScopeResolver(
-            $broker,
-            $this->getParser(),
-            $printer,
-            new FileTypeMapper($this->getParser(), $phpDocStringResolver, $this->createMock(Cache::class)),
-            new FileHelper('/'),
-            true,
-            true,
-            [
-                \EarlyTermination\Foo::class => [
-                    'doFoo',
-                ],
-            ]
-        );
-        $broker = $this->createBroker(
-            $dynamicMethodReturnTypeExtensions, 
-            $dynamicStaticMethodReturnTypeExtensions
-        );
-
-        // NOTE(Jake): 2018-04-22
-        //
-        // Hack in DynamicFunctionReturnType support
-        // -DUPLICATE CODE-
-        //
-        if ($dynamicFunctionReturnTypeExtensions) {
-            $hack = $broker->getDynamicFunctionReturnTypeExtensions();
-            $hack = array_merge($hack, $dynamicFunctionReturnTypeExtensions);
-            foreach ($dynamicFunctionReturnTypeExtensions as $extension) {
-                if ($extension instanceof BrokerAwareExtension) {
-                    $extension->setBroker($broker);
-                }
-            }
-            $refProperty = new \ReflectionProperty($broker, 'dynamicFunctionReturnTypeExtensions');
-            $refProperty->setAccessible(true);
-            $refProperty->setValue($broker, $hack);
-        }
-        
-        $resolver->processNodes(
-            $this->getParser()->parseFile($file),
-            new Scope(
-                $broker,
-                $printer,
-                new TypeSpecifier($printer),
-                $file
-            ),
-            $callback
-        );
-    }
-
-    private function assertTypeDescribe(string $expectedDescription, string $actualDescription, string $label = '')
-    {
-        $this->assertSame(
-            $expectedDescription,
-            $actualDescription,
-            $label
         );
     }
 }
